pyproject.toml
[project]
name = "f1-race-predition"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastf1==3.3.9",
    "imbalanced-learn==0.12.3",
    "ipykernel==6.29.4",
    "joblib==1.4.2",
    "jupyter==1.0.0",
    "matplotlib==3.9.0",
    "numpy==1.26.4",
    "pandas==2.2.2",
    "psycopg2-binary==2.9.9",
    "pyarrow==16.1.0",
    "pytest==8.2.1",
    "pytest-cov==5.0.0",
    "python-dotenv==1.0.1",
    "pyyaml==6.0.1",
    "requests==2.31.0",
    "scikit-learn==1.5.0",
    "seaborn==0.13.2",
    "shap==0.45.1",
    "sqlalchemy==2.0.30",
    "streamlit==1.35.0",
    "tqdm==4.66.4",
    "xgboost==2.0.3",
]
